requirements.txt
streamlit==1.31.0
pandas==2.1.4
pydeck==0.8.0
python-dotenv==1.0.0
google-generativeai==0.3.2
openai==1.12.0
requests==2.31.0
pillow==10.2.0
geopy==2.4.1
folium==0.15.0
streamlit-folium==0.15.0 
aiohttp==3.11.13